requirements.txt
django==2.1
djangorestframework==3.8.2
django-cors-headers==2.4.0
psycopg2
